requirements.txt
Django==2.2.5
django-extensions==2.2.1
gunicorn==19.9.0
djangorestframework==3.10.3
dj-static==0.0.6